WIP: Revise element style updates process
#changelog

CSS_cache is now the primary mechanism for updating a component element.
When a component element is selected, a CSS_cache object is attached to
the selection object. This is a departure from creating a CSS_cache object
only when an action is initiated. This allows the system to pre-cache
the necessary information to update the element.

This also allows us to delay sending update data back to server until
the element is deselected or some other condition that forces
synchronization is met.

When synchronizing, a group of mutating actions can be combined into a
change list allowing more verbose change descriptions to be sent back to
server. This means a transition from one component to another can include
multiple changes to one or more elements and CSS rules.